Python/nasa_apod/requirements.txt
Flask==2.0.3
Pillow==9.4.0
requests==2.27.1
